added super-admin-utils class
