+++ core/tests/puppeteer-acceptance-tests/puppeteer-testing-utilities/puppeteer-utils.js
@@ -18,14 +18,8 @@
 
 const puppeteer = require('puppeteer');
 const testConstants = require('./test-constants.js');
-const { showMessage } = require('./show-message-utils.js');
 
-const rolesEditorTab = testConstants.URLs.RolesEditorTab;
-const roleEditorInputField = 'input.e2e-test-username-for-role-editor';
-const roleEditorButtonSelector = 'button.e2e-test-role-edit-button';
-const rolesSelectDropdown = 'div.mat-select-trigger';
 const LABEL_FOR_SUBMIT_BUTTON = 'Submit and start contributing';
-const addRoleButton = 'button.oppia-add-role-button';
 
 module.exports = class puppeteerUtilities {
   constructor() {
@@ -49,7 +43,7 @@ module.exports = class puppeteerUtilities {
         this.page = await browser.newPage();
         await this.page.setViewport({ width: 0, height: 0 });
         // Accepting the alerts that appear in between the tests.
-        await this.page.on('dialog', async dialog => {
+        this.page.on('dialog', async(dialog) => {
           await dialog.accept();
         });
       });
@@ -80,6 +74,9 @@ module.exports = class puppeteerUtilities {
    */
   async signUpNewUser(userName, signInEmail) {
     await this.signInWithEmail(signInEmail);
+    if (this.page.url() === testConstants.URLs.LearnerDashboard) {
+      return;
+    }
     await this.type('input.e2e-test-username-input', userName);
     await this.clickOn('input.e2e-test-agree-to-terms-checkbox');
     await this.page.waitForSelector(
@@ -109,78 +106,6 @@ module.exports = class puppeteerUtilities {
       await this.page.waitForSelector(selector);
       await this.page.click(selector);
     }
-  }
-
-  /**
-   * The function to assign a role to a user.
-   * @param {string} username - The username to which role would be assigned.
-   * @param {string} role - The role that would be assigned to the user.
-   */
-  async assignRoleToUser(username, role) {
-    await this.goto(rolesEditorTab);
-    await this.type(roleEditorInputField, username);
-    await this.clickOn(roleEditorButtonSelector);
-    await this.clickOn(addRoleButton);
-    await this.clickOn(rolesSelectDropdown);
-    await this.page.evaluate(async(role) => {
-      const allRoles = document.getElementsByClassName('mat-option-text');
-      for (let i = 0; i < allRoles.length; i++) {
-        if (allRoles[i].innerText.toLowerCase() === role) {
-          allRoles[i].click({waitUntil: 'networkidle0'});
-          return;
-        }
-      }
-      throw new Error(`Role ${role} does not exists.`);
-    }, role);
-  }
-
-  /**
-   * The function excepts the user to have the given role.
-   * @param {string} username - The username to which role must be assigned.
-   * @param {string} role - The role which must be assigned to the user.
-   */
-  async expectUserToHaveRole(username, role) {
-    const currentPageUrl = this.page.url();
-    await this.goto(rolesEditorTab);
-    await this.type(roleEditorInputField, username);
-    await this.clickOn(roleEditorButtonSelector);
-    await this.page.waitForSelector('div.justify-content-between');
-    await this.page.evaluate((role) => {
-      const userRoles = document.getElementsByClassName(
-        'oppia-user-role-description');
-      for (let i = 0; i < userRoles.length; i++) {
-        if (userRoles[i].innerText.toLowerCase() === role) {
-          return;
-        }
-      }
-      throw new Error(`User does not have the ${role} role!`);
-    }, role);
-    showMessage(`User ${username} has the ${role} role!`);
-    await this.goto(currentPageUrl);
-  }
-
-  /**
-   * The function excepts the user to not have the given role.
-   * @param {string} username - The user to which the role must not be assigned.
-   * @param {string} role - The role which must not be assigned to the user.
-   */
-  async expectUserNotToHaveRole(username, role) {
-    const currentPageUrl = this.page.url();
-    await this.goto(rolesEditorTab);
-    await this.type(roleEditorInputField, username);
-    await this.clickOn(roleEditorButtonSelector);
-    await this.page.waitForSelector('div.justify-content-between');
-    await this.page.evaluate((role) => {
-      const userRoles = document.getElementsByClassName(
-        'oppia-user-role-description');
-      for (let i = 0; i < userRoles.length; i++) {
-        if (userRoles[i].innerText.toLowerCase() === role) {
-          throw new Error(`User has the ${role} role!`);
-        }
-      }
-    }, role);
-    showMessage(`User ${username} does not have the ${role} role!`);
-    await this.goto(currentPageUrl);
   }
 
   /**
