+++ core/tests/puppeteer-acceptance-tests/puppeteer-testing-utilities/test-constants.js
@@ -19,6 +19,7 @@
 let testConstants = {
   URLs: {
     home: 'http://localhost:8181/',
+    LearnerDashboard: 'http://localhost:8181/learner-dashboard',
     BlogDashboard: 'http://localhost:8181/blog-dashboard',
     BlogAdmin: 'http://localhost:8181/blog-admin',
     CreatorDashboard: 'http://localhost:8181/creator-dashboard',
